+++ inst/assets/sw-dropdown/sw-dropdown.js
@@ -83,6 +83,12 @@ function swDrop(
       ) && // hack for bootstrap select
       !(
         $(e.target).parents('.bootstrap-select').length
+      ) && // hack for virtual select
+      !(
+        $(e.target).parents('.vscomp-value-tag').length
+      ) && // hack for virtual select
+      !(
+        $(e.target).parents('.vscomp-value-tag-clear-button').length
       ) && // hack for pickr color
       !(
         $(e.target).parents('.pcr-app').length