+++ NEWS.md
@@ -1,3 +1,14 @@
+shinyWidgets 0.9.0
+======================
+
+* New widget `calendarProInput()`, a date and time picker component based on JavaScript library [vanilla-calendar-pro](https://github.com/uvarov-frontend/vanilla-calendar-pro).
+* `virtualSelectInput()`: ability to set a JavaScript function as callback for `selectedLabelRenderer`.
+* New argument `inputType` in `searchInput()` and `textInputIcon()` to specify the input's type, for example `"password"`.
+* `knobInput()` and `switchInput()` are softly deprecated, since the JavaScript libraries used by those widgets are no longer actively maintained.
+* Updated virtual-select to 1.0.47 (fix focus issue), see https://github.com/sa-si-dev/virtual-select/releases.
+* Updated @simonwep/pickr to 1.9.1, see https://github.com/simonwep/pickr/releases.
+
+
 shinyWidgets 0.8.7
 ======================